Add files via upload
Introduction

Face Recognition is a system of identifying and verifying people. This application finds various uses in modern world.
The 'face_data_collect.py' is used to collect facial data of people, which is stored in 'data' repository. 
'face_recognition.py' inputs video stream from camera and recognize people in the frame. 

For information in depth please go through readme.txt and 'Amril_Prasad(43913013).pdf' which presents a research paper in IEEE format.